Added an argument to customize the display format of the composition ratio
The following modification is an example of adding an argument (e.g. label_ratio_format) to drawTreemap.R to customize the display format of the composition ratio and passing it to draw_label_voronoi() and draw_label_sunburst(). Also update the documentation so that users can understand the new functionality.